Rename "dwelltime" to "dwell_time" in filenames
